+++ tests/testthat/test-backend-postgres.R
@@ -17,6 +17,8 @@ test_that("custom stringr functions translated correctly", {
 
   expect_equal(translate_sql(str_detect(x, y)), sql("`x` ~ `y`"))
   expect_equal(translate_sql(str_detect(x, y, negate = TRUE)), sql("!(`x` ~ `y`)"))
+  expect_equal(translate_sql(str_like(x, y)), sql("`x` ILIKE `y`"))
+  expect_equal(translate_sql(str_like(x, y, FALSE)), sql("`x` LIKE `y`"))
   expect_equal(translate_sql(str_replace(x, y, z)), sql("REGEXP_REPLACE(`x`, `y`, `z`)"))
   expect_equal(translate_sql(str_replace_all(x, y, z)), sql("REGEXP_REPLACE(`x`, `y`, `z`, 'g')"))
   expect_equal(translate_sql(str_squish(x)), sql("LTRIM(RTRIM(REGEXP_REPLACE(`x`, '\\s+', ' ', 'g')))"))