+++ NEWS.md
@@ -1,5 +1,8 @@
 # dbplyr (development version)
 
+* `stringr::str_like()` (new in 1.5.0) is translated to the closest `LIKE` 
+  equivalent (@rjpat, #509)
+
 * Postgres backend generates correct literals for Dates (#727).
 
 * `NA`s in `blob()`s are correctly translated to `NULL` (#983).